pyproject.toml
[tool.poetry]
name = "dbpool"
version = "1.1.0"
description = "Enhance mysql-connector-python pooling"
authors = ["liyong <hungrybirder@gmail.com>"]
license = "MIT"
homepage = "https://github.com/hungrybirder/dbpool-python"
classifiers = [
    "Environment :: Console",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10"
]

[tool.poetry.dependencies]
python = ">=3.6,<4"
mysql-connector-python = "^8.0.27"

[tool.poetry.dev-dependencies]
tox = "^3.24.5"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0", "wheel", "setuptools>50"]
build-backend = "poetry.core.masonry.api"
